Implement pre-sync
The pre-sync reduces the memory usage of the synchronization by first shift the
trace to the origin, such that if they were started at the same time, they
should overlap quickly. This yield a first approximation of the
synchronization. The complete sync is performed, but then events are matched
continuously, thus reducing the maximum memory consumption.

step        hit   max size
worst     18884     124835
presync    1404      10317
realsync  13853      10758

The realsync also includes cleanup of old unmached packets, and that explain
the reduced max size.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>