Test for bidirectional equality
If setting a timestamp transform b that is a subclass of a, the following
holds:

a.equals(b) == true
b.equals(a) == false

This is caused by the operator instanceof. Test that two objects are really
equals.

Not sure if this is the right fix for this problem. What is the best practice
to test equality among class hierarchy?

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>